Making progress on issue view
I've changed all querys so now all data is got with a single query and
then stored in IndexedDB. This would make it load fast. Also working on
displaying all data about issue. Pull request view will have to wait a
bit but when issue view is finished it'll be easy to make pull request
view since most of data is the same. There should be a component for
each type of data so it'll be easy to make other views.